split out extensiontests
